Move periods_test.go into dataset_test.go
Having a separate file wasn’t being counted in coverage statistics (or
even being run?).
